Crate edn

Re-exports

pub use intern_set::InternSet;

pub use value_rc::Cloned;

pub use value_rc::FromRc;

pub use value_rc::ValueRc;

pub use parse::ParseError;

pub use types::FromMicros;

pub use types::FromMillis;

pub use types::Span;

pub use types::SpannedValue;

pub use types::ToMicros;

pub use types::ToMillis;

pub use types::Value;

pub use types::ValueAndSpan;

pub use symbols::Keyword;

pub use symbols::NamespacedSymbol;

pub use symbols::PlainSymbol;

Modules

entities
This module defines core types that support the transaction processor.

intern_set

matcher

parse

pretty_print

query

symbols

types

utils

value_rc

Macros

ns_keyword

Structs

BigInt
A big signed integer type.

DateTime
ISO 8601 combined date and time with time zone.

OrderedFloat
A wrapper around Floats providing an implementation of Ord and Hash.

Utc
The UTC time zone. This is the most efficient time zone when you don't need the local time. It is also used as an offset (which is also a dummy type).

Uuid
A Universally Unique Identifier (UUID).

Enums

UuidParseError
Error details for string parsing failures.