Trait edn::entities::TransactableValueMarker

pub trait TransactableValueMarker { }

EntityPlace and ValuePlace embed values, either directly (i.e., ValuePlace::Atom) or indirectly (i.e., EntityPlace::LookupRef). In order to maintain the graph of Into and From relations, we need to ensure that {Value,Entity}Place can't match as a potential value. (If it does, the impl Into<T> for T default conflicts.) This marker trait allows to mark acceptable values, thereby removing {Entity,Value}Place from consideration.