Struct edn::intern_set::InternSet

pub struct InternSet<T> where
    T: Eq + Hash,  { /* fields omitted */ }

An InternSet allows to "intern" some potentially large values, maintaining a single value instance owned by the InternSet and leaving consumers with lightweight ref-counted handles to the large owned value. This can avoid expensive clone() operations.

In Mentat, such large values might be strings or arbitrary [a v] pairs.

See https://en.wikipedia.org/wiki/String_interning for discussion.

Methods

impl<T> InternSet<T> where
T: Eq + Hash,

pub fn new() -> InternSet<T>

pub fn intern<R: Into<ValueRc<T>>>(&mut self, value: R) -> ValueRc<T>

Intern a value, providing a ref-counted handle to the interned value.

use edn::{InternSet, ValueRc};

let mut s = InternSet::new();

let one = "foo".to_string();
let two = ValueRc::new("foo".to_string());

let out_one = s.intern(one);
assert_eq!(out_one, two);
// assert!(!&out_one.ptr_eq(&two));      // Nightly-only.

let out_two = s.intern(two);
assert_eq!(out_one, out_two);
assert_eq!(1, s.len());
// assert!(&out_one.ptr_eq(&out_two));   // Nightly-only.