Module edn::query

Re-exports

pub use Keyword;

pub use PlainSymbol;

Structs

Aggregate

NamedPullAttribute

NotJoin

OrJoin

Order
An abstract declaration of ordering: direction and variable.

ParsedQuery

Pattern

Predicate

Pull

QueryFunction

TypeAnnotation

Variable

WhereFn

Enums

Binding

Direction

Element

FindSpec
A definition of the first part of a find query: the [:find ?foo ?bar…] bit.

FnArg

IdentOrEntid

Limit

NonIntegerConstant
These are the scalar values representable in EDN.

OrWhereClause

PatternNonValuePlace
e, a, tx can't be values -- no strings, no floats -- and so they can only be variables, entity IDs, ident keywords, or placeholders. This encoding allows us to represent integers that aren't entity IDs. That'll get filtered out in the context of the database.

PatternValuePlace
The v part of a pattern can be much broader: it can represent integers that aren't entity IDs (particularly negative integers), strings, and all the rest. We group those under Constant.

PullAttributeSpec

PullConcreteAttribute

SrcVar

UnifyVars

VariableOrPlaceholder

WhereClause

Traits

ContainsVariables

FromValue

Type Definitions

SrcVarName