Enum edn::types::SpannedValue

pub enum SpannedValue {
    Nil,
    Boolean(bool),
    Integer(i64),
    Instant(DateTime<Utc>),
    BigInteger(BigInt),
    Float(OrderedFloat<f64>),
    Text(String),
    Uuid(Uuid),
    PlainSymbol(PlainSymbol),
    NamespacedSymbol(NamespacedSymbol),
    Keyword(Keyword),
    Vector(Vec<ValueAndSpan>),
    List(LinkedList<ValueAndSpan>),
    Set(BTreeSet<ValueAndSpan>),
    Map(BTreeMap<ValueAndSpan, ValueAndSpan>),
}

SpannedValue is the parallel to Value but used in ValueAndSpan. Container types have ValueAndSpan children.

Variants
NilBoolean(bool)Integer(i64)Instant(DateTime<Utc>)BigInteger(BigInt)Float(OrderedFloat<f64>)Text(String)Uuid(Uuid)PlainSymbol(PlainSymbol)NamespacedSymbol(NamespacedSymbol)Keyword(Keyword)Vector(Vec<ValueAndSpan>)List(LinkedList<ValueAndSpan>)Set(BTreeSet<ValueAndSpan>)Map(BTreeMap<ValueAndSpan, ValueAndSpan>)
Methods

impl SpannedValue

pub fn is_nil(&self) -> bool

pub fn is_boolean(&self) -> bool

pub fn is_integer(&self) -> bool

pub fn is_instant(&self) -> bool

pub fn is_big_integer(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_text(&self) -> bool

pub fn is_uuid(&self) -> bool

pub fn is_symbol(&self) -> bool

pub fn is_namespaced_symbol(&self) -> bool

pub fn is_vector(&self) -> bool

pub fn is_list(&self) -> bool

pub fn is_set(&self) -> bool

pub fn is_map(&self) -> bool

pub fn is_keyword(&self) -> bool

pub fn is_namespaced_keyword(&self) -> bool

pub fn as_nil(&self) -> Option<()>

as_nil does not use the macro as it does not have an underlying value, and returns Option<()>.

pub fn as_boolean(&self) -> Option<bool>

pub fn as_integer(&self) -> Option<i64>

pub fn as_instant(&self) -> Option<DateTime<Utc>>

pub fn as_float(&self) -> Option<f64>

pub fn as_big_integer(&self) -> Option<&BigInt>

pub fn as_ordered_float(&self) -> Option<&OrderedFloat<f64>>

pub fn as_text(&self) -> Option<&String>

pub fn as_uuid(&self) -> Option<&Uuid>

pub fn as_symbol(&self) -> Option<&PlainSymbol>

pub fn as_namespaced_symbol(&self) -> Option<&NamespacedSymbol>

pub fn as_keyword(&self) -> Option<&Keyword>

pub fn as_plain_keyword(&self) -> Option<&Keyword>

pub fn as_namespaced_keyword(&self) -> Option<&Keyword>

pub fn as_vector(&self) -> Option<&Vec<ValueAndSpan>>

pub fn as_list(&self) -> Option<&LinkedList<ValueAndSpan>>

pub fn as_set(&self) -> Option<&BTreeSet<ValueAndSpan>>

pub fn as_map(&self) -> Option<&BTreeMap<ValueAndSpan, ValueAndSpan>>

pub fn into_boolean(self) -> Option<bool>

pub fn into_integer(self) -> Option<i64>

pub fn into_instant(self) -> Option<DateTime<Utc>>

pub fn into_big_integer(self) -> Option<BigInt>

pub fn into_ordered_float(self) -> Option<OrderedFloat<f64>>

pub fn into_float(self) -> Option<f64>

pub fn into_text(self) -> Option<String>

pub fn into_uuid(self) -> Option<Uuid>

pub fn into_symbol(self) -> Option<PlainSymbol>

pub fn into_namespaced_symbol(self) -> Option<NamespacedSymbol>

pub fn into_keyword(self) -> Option<Keyword>

pub fn into_plain_keyword(self) -> Option<Keyword>

pub fn into_namespaced_keyword(self) -> Option<Keyword>

pub fn into_vector(self) -> Option<Vec<ValueAndSpan>>

pub fn into_list(self) -> Option<LinkedList<ValueAndSpan>>

pub fn into_set(self) -> Option<BTreeSet<ValueAndSpan>>

pub fn into_map(self) -> Option<BTreeMap<ValueAndSpan, ValueAndSpan>>

pub fn from_bigint(src: &str) -> Option<SpannedValue>

pub fn from_float(src: f64) -> SpannedValue

pub fn from_ordered_float(src: OrderedFloat<f64>) -> SpannedValue

pub fn from_symbol<'a, T: Into<Option<&'a str>>>(
namespace: T,
name: &str
) -> SpannedValue

pub fn from_keyword<'a, T: Into<Option<&'a str>>>(
namespace: T,
name: &str
) -> SpannedValue

pub fn is_collection(&self) -> bool

pub fn is_atom(&self) -> bool

pub fn into_atom(self) -> Option<SpannedValue>