Enum edn::query::Element

pub enum Element {
    Variable(Variable),
    Aggregate(Aggregate),
    Corresponding(Variable),
    Pull(Pull),
}

Variants
Variable(Variable)Aggregate(Aggregate)Corresponding(Variable)

In a query with a max or min aggregate, a corresponding variable (indicated in the query with (the ?var), is guaranteed to come from the row that provided the max or min value. Queries with more than one max or min cannot yield predictable behavior, and will err during algebrizing.

Pull(Pull)
Methods

impl Element

pub fn is_unit(&self) -> bool

Returns true if the element must yield only one value.